requirements.txt
groq
deepgram-sdk
python-dotenv
websockets
